Migrate config to TypeScript
