+++ src/types/index.ts
@@ -6,5 +6,6 @@ export { default as ICommandConfig } from "./ICommandConfig";
 export { default as IKomvosEvent } from "./IKomvosEvent";
 
 // Types
+export { default as KomvosConfig } from "./KomvosConfig";
 export { default as KomvosClient } from "./KomvosClient";
 export { default as Settings } from "./Settings";
